requirements.txt
fastapi
uvicorn
python-docx
openpyxl
sqlalchemy
numpy
mysql-connector-python
docxtpl
pandas
pymysql
mariadb
python-dotenv

